Add auto start to bind when the wan interface goes up.
A problem with this change is that bind comes up before the firewall does
in some cases.

It would be better to have it come up after the firewall, but there is
no room in the current naming scheme for this to do so.

A better patch would be to change to:

20-firewall # start the firewall
21-ntpdate or rdate AGAINST AN IP address -
22-bind
23-ntpd - If the current time is wrong fix it somehow