The IPv6 allowable DNS clients was overly broad
Allowing the entire world access to your internal DNS view is bad.
Not allowing yourself access to internal DNS via ipv6 is also bad.

A better solution would be to generate this acl on connection to
the internet, after an external 6to4 address has been calculated.

But for now, this will require hand editing.