Congestion control research sets buffers = 50
And rather than fight to obtain to so as re-run those models with
current (insane) linux defaults for buffering - setup so that
we are running close to that throughout.

So txqueuelen 40, and driver buffers of 8 is hopefully
'close enough' to more closely match theoretical results.

That said, I was pretty happy with driver buffers of 4,
and txqueuelen of 8 - on ethernet... and THAT said,
I hope to switch to QFQ in the near future so a little
more buffering is ok.

Also, a problem was noted on wireless with odd numbers
for txqueuelen (formerly 37) which did not occur with
an even number.