Updated debloat script to the one in the deBloat repo
No longer do we modify txqueuelen, but enable sfq on all
interfaces with an equivalent packet limit (40), to what
we were using for txqueuelen.

While this is a decent idea for ethernet, it
will have side effects on packet aggregation.

The debloat script needs a conf file format designed.

My principal purpose in this is to thoroughly exercise the
new qdiscs, and also make sure that the interfaces STAY
modified with the new qdiscs when brought up and down.