Renamed polipo-latest back to polipo
When I'm not explicitly trying to push the bleeding edge and
maintain two versions to play with, it makes more sense to
use the openwrt feed preference function when installing
the package. This makes for minimal differences between the two
and an easy update to push out to openwrt head when it proves
stable.