SFQ is limited in packets. RED is in bytes. tcrules.awk conflated these
The original openwrt shaper took the RED byte calculation...
and reused it to specify a limit for sfq. However, sfq uses
packets rather than bytes, so it was specifying, say 16000 bytes
and translating that to 16000 packets.

This was not an error I prior to 3.3, because SFQ had a
hard coded limit of 127 packets. It is now.

So this commit puts a lower and upper bound on the maximum packets
that is sane, but is not pre 3.2 compatible.