Attempt to cache ALL the roots dynamically with zone transfers
Previously in cerowrt, we did a zone transfer from f only.
In india, f is reachable rarely.

There are two choices: use db.hints or zone transfers

1) rely on the db.root hints distribution.

2) attempt to cache all the roots via zone transfers

At the moment I choose 2 - because keeping db.root updated
involves cron, wget, checksumming, and restarting the server.

Huge debate on this on the mailing list, but I'd like to cache
as much as possible.