named: caching the roots has too many dependencies
While the performance benefit of caching the roots seems good...

on boot, this leads to cascade failures while dnssec is not working.
bind ends up throttling the connections to the roots, or vice versa
and it can take minutes to actually get anywhere, during which time
ntp is kicking the nameserver and vice versa, and sometimes
we never get anywhere.