Relentless TCP Matt Mathis, 2008
Relentless TCP is designed to maintain a standing queue at some network
bottleneck.  It does so without regards to other traffic, and thus it is not
generally fair to other network users.  It MUST NOT be used unless the network
is designed to actively control it to protect other flows and the Internet at
large from its aggressiveness.

To protect other flows, the network carring Relentless TCP traffic has to
implement Scavenger Service, Fair Queuing, Approximate Fair Dropping, or some
other capacity allocation algorithm.

To protect the Internet at large the network carring Relentless TCP traffic
has to be a physically or logically isolated from the rest of the Internet.
Be espically warry of the potential for dynamic routing to choose an
alternative path that can not adiquiately contol Relentless TCP.

To minimize the risks associated with Relentless TCP, it has to be enabled on a
per-connection basis.