Added babelm branch which contains a new smoothed metric algo
I note that this experimental package also contains a patch
to mark babel packets as ECN capable, when in reality they
are not, at present.

In the words of Juliusz:

"that's a horrible kludge.  You're applying AQM to locally-originated data in
order to compensate for the lack of backpressure, and then falsely
asserting ECN in order to bypass the AQM policy."

As for the first statement, queues are necessary at multiple layers
in the stack, so I can think of no effective means of supplying
backpressure at insertion time that will work. I can think of means of
supplying congestion related drop indications to babel, but that is as
yet, unimplemented.

As to the second objection, yes this is bypassing AQM policy, however,
at some point, using ECN marking in babel could be used for something.

The udp markings can be easily obtained with pktinfo in userspace,
and I do need to produce a patch for that.

In the interim, the problem I am trying to solve is that when a rate
change happens in wifi, it is abrupt and can take a long time (with
the fq_codel aqm rapidly ramping up packet drop) to drop the backlog back to
a normal level for the new rate. IMHO it is better to have a radio
continue to function rather that potentially stop entirely because it
dropped a bunch of babel packets while seeking a new equilibrium.