quagga: add config notes for redistribute kernel and rxcost
I use the rxcost parameter to give a mild bias to the 5.x
ghz radios. In the test deployment I found it necessary to bump up
rxcost in proportion to the radio's quality and intent.

so I ended up with rxcost of 1024 and 1023 for wndr3800s
rxcost of 768 for picostation 2HPs (used by users)
rxcost 256 for the directional nanostation M5 backbone

This gave the m5 radios preference and only in case of failure
do things fall back to the 2HPs and wndr3800s. Otherwise
the m5s would prefer a direct connection to a lousy radio
vs a 2 hop or more route that used good radios and ethernet.

I'm not saying this is correct, but it worked.

redistribute kernel works correctly for dhcpv4 inserted default
routes. It does not appear to work correctly for RA inserted
ipv6 routes, nor has it been tested for other forms of ipv6 routes
as yet. WIP.