uftp4: update to head
The protocol.txt file lists the updated packet formats and the
details of how Elliptic Curve crypto is integrated.  The -k and -K
options on both sides have changed a bit to support Elliptic Curve.
I also added a -r option to specify an initial, min, and max GRTT,
and a -s option to set the robustness factor, which is how many
times the server will resend a particular message before giving up.

Those two should add some level of control that was lost with the
removal of the old timing parameters.

(I am uncertain if elliptic curve is available in openwrt's
crypto implementation)