aqm-scripts: Fix simplest.qos DOWNLINK
I goofed in making simplest.qos as simple as possible in that I
leveraged the CEIL variable, which no longer exists, and worse,
only tested it in a symmetric scenario which didn't expose the
problem.

Also r2q is just plain wrong and needs to be calculated correctly
as per

http://www.mail-archive.com/lartc@mailman.ds9a.nl/msg13343.html

Bu the math makes my head hurt and I'd rather aim for the lowest
latency/best mixing default rather than one that was computationally
cheap, because it's computationally cheap anyway.

So haven't fixed that yet.