aqm-scripts:
get_r2q was too hard to deal with. fix the htb rate at the quantum for now.
Might have to tweak that at higher rates.

add support for autotuning the number of flows
add support for autotuning ecn on or off globally
add some saner support for detecting the underlying mtu