aqm-scripts: make ecn separately configurable for ingress and egress
At the higher rates and RTTs I'm fiddling with, it makes sense to
be able to clearly see what ecn or lack thereof in either direction
can do. So this add support for the ALLECN variable, which has similar
syntax to the sysctl equivalent.

ALLECN=0 # disable ecn in the qdisc entirely
ALLECN=1 # enable in both directions
ALLECN=2 # enable in the downlink only (default)