Reorganize the link layer adaptation tab
The link layer adaptation GUI was a bit distracting with too many exposed options.
Now there is one drop down box to select the link layer (none, ethernet, or atm) with
a small hint when to pick which option. The selection of the link layer adaptation mechanism
moved to the end of the advanced configuration options, since no user ever should need to set those...