sqm-scripts - raft of problems fixed
if you don't use a linklayer ("none") option busted
switched to decimal kibits
added less "slop at the top" in simple.qos
it didn't switch qdiscs
it didn't use a different qos script
a couple other fixes

A big problem remains in that save from the web browser does not
actually reload the script. There is some option to tell the
web interface to do this (and I didn't check before if this worked), that there
are dependencies that need to be executed.