SQM: Unchecking the Enable checkbox in the GUI now actually disables SQm on the selected interface
The user can no stop and tear down the qdiscs that the SQM system
installed for a given interface. To allow this functionality
SQM mintains a file called /var/run/SQM/SQM_active_on_${IFACE}
for each interface SQM is active on. Upon stopping SQM on an interface
this file gets deleted (so the state is stored in the presence of the file)
This should make it possible to specify SQM on multiple interfaces
via the GUI.