Initial changes to squash inbound diffserv and have simpler inbound
1) By default dscp values cannot be trusted e2e as multiple
providers change them in transit. So the correct "Default" for sqm
should be to reset all inbound diffserv marked values to
BE, and then reclassify.

2) It looks like the inbound rules for simple.qos were always
broken - see iptables -x -v -L for more details

3) IF you are not going to pay attention to diffserv,
it makes no sense to have a 3 tier system for inbound,
so this grabs simplest.qos's scheme and uses that if SQUASH_DSCP
is set.

4) Haven't tried updating the gui yet.

5) Bug:

It still looks like the tc based rules for diffserv *never* worked,
and I'm not sure which would be faster - iptables or diffserv
for outbound, so I wanted to fiddle with diffserv on that
and ran out of time - the code needs to use different priorities
((prio X)+1)) to work right in functions.sh