Rationalize the handling of ifb devices
The GUI currently can only create and tear down one SQM instance.
The limiting factor is that ifb devices are hard coded in run.sh,
in a quite inflexible way. This commit creates functions in functions.sh
that for a given interface will try to find an already existing ifb
defined for that device for tear down or re-use. Alternatively
the code will try to offer a currently unused IFB that has an
index <= numifbs. Work in progress.
Needs scrutiny!