Remove IFB device selection from run.sh
Selecting which IFB to use for SQM on a given interface was handeled
statically in run.sh until now. This worked well for single IFB instances
but has two short commings. a) it does not honor people using theses IFB
before calling SQM and b) it does not play well with multiple SQM instances
as it assumes IFB device num to depend on the position in the SQM config file.
And that breaks if say the second of three instances gets disabled and/or removed.
It turns out that functions.sh and stop.sh are much better suited to
take care of selecting IFB numbers and have been changed to actually do this.
Needs eyeballs and testing...