debloat: the impact on stability is not worth the cost on wireless
fiddling with the wireless qlens DOES do nice things on some
benchmarks, however, as we are trying to get stability here,
and this may be a cause of the problems, nuke the change.

Also bump the wired quantum up to 1514 from the current 1000,
which does little good. 300 might be better, but, what the heck.