Improve SQM's generation of nemad IFBs
Some version of iproute2's ip binary require an additional "name"
in the invocation used to create new named IFB devices, while older versions
do not require this string (not that it matters much, but the man page also
marks that string as optional). But optional means it should not hurt if it is there.
Tested with cerowrt 3.10.50-1.