Fix stop.sh to avoid creating spurious ifb4ge00 interfaces
stop.sh was sourcing functions.sh without IFACE being created.
In that case functions.sh was defaulting to ge00, since stop.sh was
changed to take the interface name in its first argument instead
of as an environmet variable, a spurious ifb4ge00 was created, but only
deleted if ge00 was the passed interface. For openwrt with its
ethN.N interface names that resulted in ifb4ge00 being created but
never deleted. While this probably did not cause any malfunction
it still was confusing. Hopefully it is fixed now.