[luci-app-sqm] Offer only those qdiscs to the user that have a decent chance of working
We try to detect which qdiscs are actually useable in /usr/lib/sqm/run.sh
now. We only offer those qdiscs from the master list in run.sh that do
not cause an error when attached to an ifb and fq_codel as default (which
we need as otherwise we start with an empty list and nothing will work).
Under the assumption that fq_codel is sort of mandatory in openwrt
this heuristic should work reasonably well, and at least as well as the
hard coded list we used so far... Needs testing and supporting sqm-scripts
otherwise alll that is offered is fq_codel (but for most users that is all
that is recommended anyways).

Signed-off-by: Sebastian Moeller <moeller0@gmx.de>